requirements.txt
httpx>=0.27.0
clickhouse-connect>=0.7.0
lxml>=5.0.0
python-dateutil>=2.8.0
minio>=7.2.0
